Revert "Fix multiclass for esc50"
