+++ birdset/modules/metrics/multiclass.py
@@ -2,46 +2,6 @@ import torch
 import warnings
 import torchmetrics
 
-class MulticlassMetricsConfig:
-    """
-    A class for configuring multiclass metrics used during model training and evaluation.
-
-    Attributes:
-        main_metric (Metric): The main metric used for model training.
-        val_metric_best (Metric): The metric used for model validation.
-        add_metrics (MetricCollection): A collection of additional metrics used during model training.
-        eval_complete (MetricCollection): A collection of metrics used during model evaluation.
-    """
-
-    def __init__(
-        self,
-        num_labels: int = 21,
-    ):
-        """
-        Initializes the MetricsConfig class.
-
-        Args:
-            num_labels (int): The number of labels in the dataset. Defaults to 21 as in the HSN dataset.
-        """
-        self.main_metric: torchmetrics.Metric = torchmetrics.Accuracy(
-            task="multiclass",
-            num_classes=num_labels,
-        )
-        self.val_metric_best: torchmetrics.Metric =torchmetrics.MaxMetric()
-     
-        self.add_metrics: torchmetrics.MetricCollection = torchmetrics.MetricCollection({
-            'F1': torchmetrics.F1Score(
-                task="multiclass",
-                num_classes=num_labels,
-            ),
-        })
-        self.eval_complete: torchmetrics.MetricCollection = torchmetrics.MetricCollection({
-            'acc': torchmetrics.Accuracy(
-            task="multiclass",
-            num_classes=num_labels,
-        )
-        })
-
 class BalancedAccuracy(torchmetrics.Metric):
     def __init__(self, adjusted=False, num_classes=None):
         super().__init__(dist_sync_on_step=False)
