+++ configs/experiment/local/esc50/esc50_resnet_wandb.yaml
@@ -15,9 +15,19 @@ callbacks:
   early_stopping:
     patience: 50
     min_delta: 5e-10
-  model_checkpoint:
-    monitor: "val/MulticlassAccuracy"
 
+module:
+  lr_scheduler:
+    _target_: torch.optim.lr_scheduler
+  optimizer:
+    _target_: torch.optim.Adam
+    lr: 1e-4
+    weight_decay: 0.01
+  network:
+    sampling_rate: 32000
+trainer:
+  min_epochs: 1
+  max_epochs: 100
 
 datamodule:
   loaders:
@@ -35,4 +45,3 @@ logger:
     tags: ${tags}
     group: "esc50"
     mode: online
-
