+++ configs/module/multilabel.yaml
@@ -5,6 +5,7 @@ defaults:
   - network: ast.yaml
 
 task: multilabel
+class_weights_loss: ${datamodule.dataset.class_weights_loss}
 prediction_table: False
 num_gpus: ${trainer.devices}
 
@@ -15,7 +16,7 @@ optimizer:
   weight_decay: 0.01
 
 metrics:
-  _target_: birdset.modules.metrics.multilabel.MultilabelMetricsConfig
+  _target_: birdset.modules.base_module.MetricsConfig
   num_labels: ${datamodule.dataset.n_classes}
